Fix label auto-completion not working with mouse click, closes #127.
The component AddInlineLabel loses focus when clicking on an
autocomplete option. The current text is also added as a label whenever
the input is closed. This workaround stops the input from closing while
the autocomplete is open. To prevent it from remaining stuck in the open
state, it has to close as soon as autocomplete closes.