Change data key initialization logic to be more robust.
A new data key should only be generated the very first time a user opens
the app. If it is cleared while the app is open, saving should return an
error and the user has to manually set it in settings again or reload
the page.

Previously, after clearing localStorage, the app would silently start
saving an existing database under a new data key, which could remain
undetected for quite a long time.