Fix protobufjs not generating message interface typings.
According to https://github.com/protobufjs/protobuf.js/issues/1414#issuecomment-769791397
the issue is caused by a problem with the updated jsdoc version, so it has to be manually downgraded to fix the problem.
Huge thanks to @VitalyName for the workaround!

How this has been a completely build-breaking bug for 10 months is beyond my understanding ...